#16 Initial migration to TypeScript
