+++ .npmignore
@@ -1,0 +1,11 @@
+test/
+node_modules/
+coverage/
+src/
+.github
+.nyc_output/
+.travis.yml
+appveyor.yml
+tsconfig.json
+tslint.json
+*.iml
