+++ src/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "sourceMap": true,
+    "outDir": "../lib",
+    "declaration": true
+  }
+}
+
