+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "inlineSources": false,
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "target": "ES2017"
+  }
+}
+
